requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
datetime-truncate==1.1.1
Django==4.2.9
django-admin-charts==1.3.0
django-admin-sortable2==2.1.10
django-admin-tools==0.9.3
django-admin-tools-stats==0.9.0
django-bower==5.2.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-jsonfield==1.4.1
django-light==0.1.0.post3
django-memoize==2.3.1
django-modeltranslation==0.18.11
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-qsstats-magic==1.1.0
django-rosetta==0.10.0
djangorestframework==3.14.0
djcacheutils==3.0.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pillow==10.2.0
polib==1.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-memcached==1.61
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.2.0
